// src/thor/model/MediaObject.ts
export type MediaObject  = {
    /**
     * 
     * @type {string}
     * @memberof MediaObject
     */
    sheetId?: string;
    /**
     * 
     * @type {string}
     * @memberof MediaObject
     */
    type?: MediaObjectTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof MediaObject
     */
    url?: string;
    /**
     * 
     * @type {number}
     * @memberof MediaObject
     */
    positionRow?: number;
    /**
     * 
     * @type {number}
     * @memberof MediaObject
     */
    positionCol?: number;
    /**
     * Unique identifier for object in the system
     * @type {string}
     * @memberof MediaObject
     */
    id?: string;
    /**
     * UUID of owner of the object in the system
     * @type {string}
     * @memberof MediaObject
     */
    ownerId?: string;
    /**
     * Date of object creation
     * @type {Date}
     * @memberof MediaObject
     */
    createdDate?: Date;
    /**
     * Data, including hash of the key(s) used to encrypt this record.
     * @type {string}
     * @memberof MediaObject
     */
    keyHash?: string;
    /**
     * Last user to access object
     * @type {string}
     * @memberof MediaObject
     */
    lastAccessedById?: string;
    /**
     * Timestamp of last access of object
     * @type {Date}
     * @memberof MediaObject
     */
    lastAccessedDate?: Date;
    /**
     * Unique identifier for user who last modifed the object in the system
     * @type {string}
     * @memberof MediaObject
     */
    lastModifiedById?: string;
    /**
     * Date of last object modification
     * @type {Date}
     * @memberof MediaObject
     */
    lastModifiedDate?: Date;
}
// ...
export enum MediaObjectTypeEnum {
    MEDIAIMAGE = 'image',
    MEDIAAUDIO = 'audio',
    MEDIAVIDEO = 'video',
    MEDIASHAPE = 'shape'
}